requirements.txt
# Azure Services - Use stable FormRecognizer instead of beta DocumentIntelligence
azure-ai-formrecognizer==3.3.3
openai==1.12.0
azure-identity==1.15.0

# Web Frameworks
fastapi==0.109.2
uvicorn[standard]==0.27.1
streamlit==1.31.0

# Data Processing
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Vector Store & Embeddings
chromadb==0.4.22
langchain==0.1.7
langchain-openai==0.0.5
langchain-community==0.0.20

# Utilities
requests==2.31.0
Pillow==10.2.0
pdf2image==1.17.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Logging & Monitoring
structlog==24.1.0

# Testing (Optional but Recommended)
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0